+++ routes/services/habit/createHabit.ts
@@ -1,0 +1,6 @@
+import prisma from "../../prisma";
+import { RequestHandler } from "express";
+
+const createHabit: RequestHandler = async (req, res) => {};
+
+export default createHabit;
