+++ routes/services/habit/deleteHabit.ts
@@ -1,0 +1,6 @@
+import prisma from "../../prisma";
+import { RequestHandler } from "express";
+
+const deleteHabit: RequestHandler = async (req, res) => {};
+
+export default deleteHabit;
