+++ routes/services/habit/getHabitList.ts
@@ -1,0 +1,6 @@
+import prisma from "../../prisma";
+import { RequestHandler } from "express";
+
+const getHabitList: RequestHandler = async (req, res) => {};
+
+export default getHabitList;
