+++ routes/services/habit/updateHabit.ts
@@ -1,0 +1,6 @@
+import prisma from "../../prisma";
+import { RequestHandler } from "express";
+
+const updateHabit: RequestHandler = async (req, res) => {};
+
+export default updateHabit;
