+++ routes/services/study/getStudyList.ts
@@ -1,0 +1,11 @@
+import prisma from "../../prisma";
+import { RequestHandler } from "express";
+
+const getStudyList: RequestHandler = async (req, res, next) => {
+  try {
+  } catch (error) {
+    next(error);
+  }
+};
+
+export default getStudyList;
