+++ routes/services/study/updateStudy.ts
@@ -1,0 +1,13 @@
+import prisma from "../../prisma";
+import { RequestHandler } from "express";
+import * as s from "superstruct";
+import { PatchStudy } from "../../../struct";
+
+const updateStudy: RequestHandler = async (req, res, next) => {
+  try {
+  } catch (error) {
+    next(error);
+  }
+};
+
+export default updateStudy;
